Fix Cost type (Kotlin)
Live API testing confirmed Cost is a JSON number that isn't always a
whole number (e.g. 0.09 for some destinations) — Jackson (same
library as Java) would silently truncate it to 0 when deserializing
into Int. Changed to Double. (Kotlin's ApiResponse never had a
Currency field to remove.)

+++ kotlin/src/main/kotlin/v1/models/ApiResponse.kt
@@ -10,7 +10,7 @@ class ApiResponse {
     var message: String? = null
 
     @JsonProperty(value = "Cost")
-    var cost: Int? = null
+    var cost: Double? = null
 
     @JsonProperty(value = "MsgFollowUpUniqueCode")
     var messageFollowUpCode: String? = null
